change rates_controller and add raitings to config/routes , create FactoryBot for raitings
