+++ config/routes.rb
@@ -4,6 +4,7 @@ Rails.application.routes.draw do
     namespace :v1 do
       resources :recipes do
         resources :comments, only: %i[create]
+         resources :ratings, only: %i[create]
       end
     end
   end
