+++ spec/factories/rates.rb
@@ -1,5 +1,6 @@
 FactoryBot.define do
   factory :rate do
     recipe
-    rate { Faker:
+    rate { Faker::Number.between(from: 1, to: 5) }
+  end
 end
